+++ src/kernels/Convection.jl
@@ -5,13 +5,13 @@ type Convection <: Kernel
     velocity::Vec{2, Float64}
 end
 
-@inline function computeQpResidual(kernel::Convection, qp::Int64, i::Int64)
+@inline function computeQpResidual(kernel::Convection, qp::Integer, i::Integer)
     u = kernel.u
 
     return (kernel.velocity ⋅ u.grad[qp]) * u.phi[qp][i]
 end
 
-@inline function computeQpJacobian(kernel::Convection, v::Variable, qp::Int64, i::Int64, j::Int64)::Float64
+@inline function computeQpJacobian(kernel::Convection, v::Variable, qp::Integer, i::Integer, j::Integer)::Float64
     u = kernel.u
 
     if u.id == v.id
